Fixes and Adjustments in Proton Integration
    1. Additional Compile Option: Introduced extra compile options.

    2. Explicit Path Usage: Modified add_custom_command to be invoked with an explicit path.

    3. Type Conversion Correction: Addressed a conversion between int64_t and double.

    4. Obsolete File Removal: Removed an unnecessary, manually generated file that was previously used.

Historically, I had disabled the file generation process because it was time-consuming. However, to prioritize safety and flexibility, I’ve decided to re-enable it. Consequently, each time V8 is built, a new file will be generated and utilized. Previously, I manually added the file, circumventing the need for generation during each build.

+++ v8/include/v8-platform.h
@@ -1206,7 +1206,7 @@ class Platform {
    * required.
    */
   virtual int64_t CurrentClockTimeMilliseconds() {
-    return floor(CurrentClockTimeMillis());
+    return static_cast<int64_t>(floor(CurrentClockTimeMillis()));
   }
 
   /**
